mainboard/talos-2/rom_media.c: find CBFS in PNOR
Tested on QEMU with ECC, works as long as ramstage is not comressed.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>

mainboard/talos-2/rom_media: use mmap_helper instead of own solution

This solves the issue with loading LZMA compressed ramstage. Bootblock
still fits in SEEPROM with both console and LZ4 romstage compression,
but not with verbose CBFS debug messages.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>