NOT FOR MERGE (yet): add SHA256 implementation for size calculation
Enabling vboot is not trivial. It also significantly increases the size of
bootblock. Space in SEEPROM is heavily limited.

This adds simple SHA256 implementation and is included only to check how big
impact on the binary size it has. Hopefully, after some optimisation it will
be possible to have console output, hash function and LZ4 romstage
decompression function in bootblock at the same time.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>